Fix hydration error
refs #31
